Most circuit reorganization done
