+++ src/main/java/gregicadditions/recipes/chain/FusionElementsChain.java
@@ -355,6 +355,48 @@ public class FusionElementsChain {
                 .fluidInputs(Seaborgium.getFluid(144))
                 .fluidOutputs(Vibranium.getFluid(288))
                 .buildAndRegister();
+
+        ADV_FUSION_RECIPES.recipeBuilder().duration(100).EUt(900000).coilTier(3).euStart(10000000000L).euReturn(50)
+                .fluidInputs(ScandiumTitanium50Mix.getFluid(36))
+                .fluidInputs(RadonRadiumMix.getFluid(144))
+                .fluidOutputs(MetastableHassium.getFluid(144))
+                .buildAndRegister();
+
+        ADV_FUSION_RECIPES.recipeBuilder().duration(100).EUt(75000).coilTier(2).euStart(2080000000).euReturn(40)
+                .fluidInputs(Americium.getFluid(144))
+                .fluidInputs(Titanium.getFluid(144))
+                .fluidOutputs(Tennessine.getFluid(288))
+                .buildAndRegister();
+
+        ADV_FUSION_RECIPES.recipeBuilder().duration(100).EUt(75000).coilTier(2).euStart(2080000000).euReturn(40)
+                .fluidInputs(Plutonium.getFluid(144))
+                .fluidInputs(Titanium.getFluid(144))
+                .fluidOutputs(Livermorium.getFluid(288))
+                .buildAndRegister();
+
+        ADV_FUSION_RECIPES.recipeBuilder().duration(100).EUt(75000).coilTier(2).euStart(2080000000).euReturn(40)
+                .fluidInputs(Neptunium.getMaterial().getFluid(144))
+                .fluidInputs(Titanium.getFluid(144))
+                .fluidOutputs(Moscovium.getFluid(288))
+                .buildAndRegister();
+
+        ADV_FUSION_RECIPES.recipeBuilder().duration(100).EUt(75000).coilTier(2).euStart(2080000000).euReturn(40)
+                .fluidInputs(Astatine.getFluid(144))
+                .fluidInputs(Nickel.getFluid(144))
+                .fluidOutputs(Nihonium.getFluid(288))
+                .buildAndRegister();
+
+        ADV_FUSION_RECIPES.recipeBuilder().duration(100).EUt(75000).coilTier(2).euStart(2080000000).euReturn(40)
+                .fluidInputs(Radium.getFluid(144))
+                .fluidInputs(Vanadium.getFluid(144))
+                .fluidOutputs(Roentgenium.getFluid(288))
+                .buildAndRegister();
+
+        ADV_FUSION_RECIPES.recipeBuilder().duration(100).EUt(75000).coilTier(2).euStart(2080000000).euReturn(40)
+                .fluidInputs(Polonium.getFluid(144))
+                .fluidInputs(Chrome.getFluid(144))
+                .fluidOutputs(Meitnerium.getFluid(288))
+                .buildAndRegister();
     }
 
     public static void fusionRecipes() {
